refactor: changed routes to reflect simpler module structure
