--- web-client/src/modules/settings/pages/routes/SettingsRoute/SettingsRoute.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import SettingsContainer from '../../../containers/SettingsContainer/SettingsContainer';
-
-const SettingsRoute: React.FC = () => <SettingsContainer />;
-
-export default SettingsRoute;
